refactor user file into module, organized functions in to different files of module, corrcted imports for new structure. user.py remains with functions that should be moved to a separate module or the pong app PC-82
