+++ srcs/backend/transcendence/srcs/app/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 from  . import views
-from . import user
+from .user import relations
 
 urlpatterns = [
 	path('register/', views.register_user, name="register_user"),
@@ -15,6 +15,7 @@ urlpatterns = [
 	path('home/', views.home, name="home"),
 	path('notfound/', views.notfound, name="notfound"),
 	path('manage_account/', views.manage_account, name="manage_account"),
-	path('block_user/', user.block_user, name="block_user"),
+	path('block_user/', relations.block_user, name="block_user"),
 ]
 
+# TODO block user need divverent view definition.
