+++ srcs/backend/transcendence/srcs/app/views.py
@@ -7,8 +7,9 @@ import logging
 from .models import CustomUser
 # from django.contrib.auth import get_user_model
 from django.contrib.auth import authenticate #, login, logout
-from .forms import RegistrationForm, LoginForm, DeleteAccountForm, UpdatePasswordForm, UpdateEmailForm, UpdateNameForm, UploadImageForm
-from . import user
+from .forms import DeleteAccountForm, UpdatePasswordForm, UpdateEmailForm, UpdateNameForm, UploadImageForm
+from app.user import session, account, dashboard, relations, user
+from app.user import profile as user_profile
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 
@@ -19,10 +20,10 @@ logger = logging.getLogger(__name__)
 def register_user(request):
 	if request.method == 'POST':
 		logger.debug('In register user POST')
-		response = user.registerPOST(request)
+		response = account.registerPOST(request)
 	elif request.method == 'GET':
 		logger.debug('In register user GET')
-		response = user.registerGET(request)
+		response = account.registerGET(request)
 	else:
 		response = JsonResponse({'error': "method not allowed. please use POST or GET"})
 	return response
@@ -31,9 +32,9 @@ def register_user(request):
 def login_user(request):
 	logger.debug('In login_user()')
 	if request.method == 'POST':
-		response = user.loginPOST(request)
+		response = session.loginPOST(request)
 	elif request.method == 'GET':
-		response = user.loginGET(request)
+		response = session.loginGET(request)
 	else:
 		response = JsonResponse({'error': "method not allowed. please use POST"})
 	return response
@@ -42,7 +43,7 @@ def login_user(request):
 def logout_user(request):
 	if request.method == 'POST':
 		logger.debug('In logout user')
-		response = user.logoutPOST(request)
+		response = session.logoutPOST(request)
 	else:
 		response = JsonResponse({'error': "method not allowed. please use POST"})
 	return response
@@ -51,7 +52,7 @@ def logout_user(request):
 def get_current_username(request):
 	logger.debug('In get_current_username()')
 	if request.method == 'GET':
-		username = user.get_current_usernameGET(request)
+		username = session.get_current_usernameGET(request)
 	else:
 		username = 'only GET allowed'
 	response = JsonResponse({'message': username})
@@ -60,7 +61,7 @@ def get_current_username(request):
 
 def check_login(request):
 	logger.debug('In check_login()')
-	if request.user.is_authenticated:
+	if request.is_authenticated:
 		return JsonResponse({'status': 'authenticated'})
 	else:
 		return JsonResponse({'status': 'not authenticated'})
@@ -70,7 +71,7 @@ def check_login(request):
 def manage_account(request):
 	logger.debug('In manage_account()')
 	if request.method =='POST':
-		response = user.manage_accountPOST(request)
+		response = account.manage_accountPOST(request)
 	else:
 		response = JsonResponse({'error': "method not allowed. please use POST"})
 	return response
@@ -79,21 +80,21 @@ def manage_account(request):
 # def delete_account(request):
 # 	logger.debug('In delete_account()')
 # 	if request.method == 'GET':
-# 		response = user.delete_accountGET(request)
+# 		response = delete_accountGET(request)
 # 	elif request.method == 'POST':
-# 		response = user.delete_accountPOST(request)
+# 		response = delete_accountPOST(request)
 # 	else:
 # 		response = JsonResponse({'error': "method not allowed. please use POST or GET"})
 # 	return response
 
-
+# TODO: there is no settingsPOST anywhere in the project. This view is not used at all
 @login_required
 def settings(request):
 	logger.debug('In settings()')
 	if request.method == 'GET':
 		return render(request, 'user/settings.html', {})
-	elif request.method == 'POST':
-		response = user.settingsPOST(request)
+	# elif request.method == 'POST':
+	# 	response = usettingsPOST(request)
 	else:
 		response = JsonResponse({'error': "method not allowed. please use POST or GET"})
 	return response
@@ -115,9 +116,9 @@ def play(request):
 def friends(request):
 	logger.debug('In friends()')
 	if request.method == 'GET':
-		response = user.friendsGET(request)
+		response = relations.friendsGET(request)
 	elif request.method == 'POST':
-		response = user.friendsPOST(request)
+		response = relations.friendsPOST(request)
 	else:
 		response = JsonResponse({'error': "method not allowed. please use POST or GET"})
 	return response
@@ -149,9 +150,9 @@ def profile(request, username):
 	if request.user.username == username:
 		self = True
 	if request.method == "GET":
-		friends = user.friendsContext(username, None, None)
+		friends = relations.friendsContext(username, None, None)
 		logger.debug(friends)
-		details = user.get_profile_details(username, self)
+		details = user_profile.get_profile_details(username, self)
 		context = {}
 		context["friends"] = friends #TODO get the friends context from Tuukka's friend stuff
 		context["details"] = details
